Audit and make y-axis flipping consistent
There have been lots of long-standing FIXME notes in glimpse_context.cc
that we shouldn't need to flip the Y axis of our debug geometry, and
we've long been aware that we've probably been lazy and flipped the
y-axis in odd places without careful consideration beyond visually
getting the right orientation.

The only places where we should now be flipping the y-axis are at points
where we convert from 2D coords (with y=0 at the top, extending down)
into 3D frame coordinates with +Y extending up and -Y, or vice-versa.
(E.g. when we first convert a new depth buffer into a point cloud or
if we project a point cloud back into a 2D image used for label
inference).

There's one notably awkward case due to the viewer using using OpenGL to
render to a 2D texture which will then be sampled via the ImGui::Image()
api... GL conventionally considers the bottom left of the texture as
(0,0) and more importantly glTexImage2D will, by default load
consecutive scanlines of an image starting at (0,0) (the bottom).  This
means that a lot of GL software will upload images upside-down out of
ease but you later need to flip them either while sampling or by
flipping your geometry.  ImGui::Image() assumes that your image is
upside down which is generally fine except for consistency we now need
to make sure that when we render to a texture we actually want to render
upside down so that the resulting image will be the right way up after
it's draw using ImGui::Image(). Technically we could also pass different
uv coords to Image() but it seems best to be consistent and maintain all
our GL textures with the same convention for top/bottom.